Combine `map` and `map_indexed`
