+++ src/_bootstrap.php
@@ -30,7 +30,6 @@ require __DIR__ . '/identity.php';
 require __DIR__ . '/instance_of.php';
 require __DIR__ . '/last.php';
 require __DIR__ . '/map.php';
-require __DIR__ . '/map_indexed.php';
 require __DIR__ . '/memoize.php';
 require __DIR__ . '/not.php';
 require __DIR__ . '/partial.php';
